requirements.txt
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
GPUtil==1.4.0
idna==3.4
keyboard==0.13.5
lxml==4.9.3
moviepy>=1.0.3
mutagen==1.47.0
numpy>=1.22.2
opencv_python>=4.7.0.7
Pillow>=9.3.0
psutil>=5.9.4
PyAutoGUI>=0.9.53
pygame>=2.1.2
PyGetWindow>=0.0.9
python_vlc>=3.0.16120
pywin32>=305
pywinauto>=0.6.8
PyYAML>=6.0
python-dotenv>=1.0.0
pycryptodomex==3.19.0
py_cpuinfo==9.0.0
pyreadline3>=2.1
python-dotenv==1.0.0
Pygments>=2.15.0
requests>=2.29.0
rich==13.6.0
semantic_version>=2.10.0
setuptools>=49.2.1
spotdl>=4.2.1
sponsorblock.py==0.2.3
shellingham==1.5.4
tabulate==0.9.0
tqdm>=4.64.1
typer==0.9.0
typing==3.7.4.3
validators==0.22.0 
yt-dlp==2023.10.13
WMI>=1.5.1
